Refactored session middleware and lucia access
