+++ apps/api/src/lib/lucia.ts
@@ -1,0 +1,26 @@
+import { DatabaseSessionAttributes, Lucia } from 'lucia';
+import { D1Adapter } from '@lucia-auth/adapter-sqlite';
+
+export function initializeLucia(D1: D1Database) {
+  const adapter = new D1Adapter(D1, {
+    user: 'user',
+    session: 'session',
+  });
+  return new Lucia(adapter, {
+    getSessionAttributes: (attributes) => {
+      const attr = attributes as DatabaseSessionAttributes;
+      return {
+        chainId: attr.chainId,
+      };
+    },
+  });
+}
+
+declare module 'lucia' {
+  interface Register {
+    Lucia: ReturnType<typeof initializeLucia>;
+  }
+  interface DatabaseSessionAttributes {
+    chainId: number;
+  }
+}
