+++ apps/api/src/routes/siwe/index.ts
@@ -1,84 +1,23 @@
 import { Context, Hono } from 'hono';
 import { generateNonce, SiweMessage } from 'siwe';
 
-import { DatabaseSessionAttributes, Lucia, Session, User } from 'lucia';
-import { getCookie } from 'hono/cookie';
-import { D1Adapter } from '@lucia-auth/adapter-sqlite';
+import { Lucia, Session, User } from 'lucia';
+
 import { env } from 'hono/adapter';
 import { csrf } from 'hono/csrf';
-
-export function initializeLucia(D1: D1Database) {
-  const adapter = new D1Adapter(D1, {
-    user: 'user',
-    session: 'session',
-  });
-  return new Lucia(adapter, {
-    getSessionAttributes: (attributes) => {
-      const attr = attributes as DatabaseSessionAttributes;
-      return {
-        chainId: attr.chainId,
-      };
-    },
-  });
-}
-
-declare module 'lucia' {
-  interface Register {
-    Lucia: ReturnType<typeof initializeLucia>;
-  }
-  interface DatabaseSessionAttributes {
-    chainId: number;
-  }
-}
+import { session } from '../../middleware/session';
 
 const app = new Hono<{
   Variables: {
     user: User | null;
     session: Session | null;
+    lucia: Lucia;
   };
 }>();
 
 app.use(csrf({ origin: ['real.sovereignnature.com', 'localhost'] })); //TODO: Localhost in dev
 
-app.use('*', async (c, next) => {
-  console.log('SESSION MIDDLEWARE');
-
-  const { SESSIONS_DB } = env<{ SESSIONS_DB: D1Database }>(c as Context);
-
-  const lucia = initializeLucia(SESSIONS_DB);
-
-  console.log('sessionCookieName', lucia.sessionCookieName);
-
-  const sessionId = getCookie(c, lucia.sessionCookieName) ?? null;
-  console.log('sessionId', sessionId);
-
-  if (!sessionId) {
-    c.set('user', null);
-    c.set('session', null);
-
-    console.log('NO SESSION ID');
-    return next();
-  }
-  const { session, user } = await lucia.validateSession(sessionId);
-
-  if (session && session.fresh) {
-    // use `header()` instead of `setCookie()` to avoid TS errors
-    c.header('Set-Cookie', lucia.createSessionCookie(session.id).serialize(), {
-      append: true,
-    });
-  }
-
-  if (!session) {
-    c.header('Set-Cookie', lucia.createBlankSessionCookie().serialize(), {
-      append: true,
-    });
-  }
-
-  c.set('user', user);
-  c.set('session', session);
-
-  return next();
-});
+app.use('*', session);
 
 app.get('/nonce', (c) => {
   return c.text(generateNonce());
@@ -86,7 +25,7 @@ app.get('/nonce', (c) => {
 
 app.post('/verify', async (c) => {
   const { SESSIONS_DB } = env<{ SESSIONS_DB: D1Database }>(c as Context);
-  const lucia = initializeLucia(SESSIONS_DB);
+  const lucia = c.get('lucia');
 
   const body = await c.req.json();
 
@@ -101,7 +40,7 @@ app.post('/verify', async (c) => {
 
   const { data: message } = await siweMessage.verify({
     signature: body.signature,
-    // nonce: session.nonce,
+    // nonce: session.nonce, //TODO: Add nonce to session
   });
 
   const address = message.address;
@@ -131,7 +70,7 @@ app.post('/verify', async (c) => {
   return c.text('true');
 });
 
-// get the session
+// Get the session
 app.get('/session', (c) => {
   const session = c.get('session');
 
@@ -142,12 +81,9 @@ app.get('/session', (c) => {
   return c.json({ address: session.userId, chainId: session.chainId });
 });
 
-//TODO: Add signout route
 app.get('/signout', (c) => {
   const session = c.get('session');
-
-  const { SESSIONS_DB } = env<{ SESSIONS_DB: D1Database }>(c as Context);
-  const lucia = initializeLucia(SESSIONS_DB);
+  const lucia = c.get('lucia');
 
   if (!session) {
     return c.json({ message: 'No session found' }, 404);
